requirements.txt
javalang
GitPython
langchain
langchain-openai
pydantic
openai
python-dotenv